strands-agents-notebooks/claim_inquiry_agent/requirements.txt
# Healthcare Agent Dependencies

# AWS SDK
boto3>=1.34.0

# Strands Agents
strands-agents
strands-agents-tools

# Bedrock AgentCore (for AgentCore deployment)
bedrock-agentcore>=0.1.0

# Logging and utilities
python-json-logger>=2.0.0
